+++ libsa4py/cst_transformers.py
@@ -823,3 +823,25 @@ class TypeQualifierResolver(cst.CSTTransformer):
                                  match.AnnAssign(target=match.Name(value=match.DoNotCare()),
                                                  annotation=match.SaveMatchedNode(match.DoNotCare(), "type")))['type']
 
+
+class ParametricTypeDepthReducer(cst.CSTTransformer):
+    def __init__(self, max_annot_depth=2):
+        self.max_annot_depth = max_annot_depth
+        self.max_annot = 0
+        self.current_annot_depth = 0
+
+    def visit_Subscript(self, node):
+        self.current_annot_depth += 1
+        if self.max_annot < self.current_annot_depth:
+            self.max_annot += 1
+
+    def leave_Subscript(self, original_node, updated_node):
+        self.current_annot_depth -= 1
+        return updated_node
+
+    def leave_SubscriptElement(self, original_node, updated_node):
+        if self.max_annot > self.max_annot_depth and self.current_annot_depth == self.max_annot_depth:
+            self.max_annot -= 1
+            return updated_node.with_changes(slice=cst.Index(value=cst.Name(value='Any')))
+        else:
+            return updated_node